backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic-settings>=2.2
sqlalchemy>=2.0
psycopg[binary]>=3.1
python-multipart>=0.0.9
passlib[argon2]>=1.7.4
python-jose[cryptography]>=3.3.0
